feat(refactor): refactored library structure and tests
