+++ package.json
@@ -58,5 +58,19 @@
   },
   "version": "0.0.0",
   "types": "lib/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./lib/index.js",
+      "require": "./lib/index.js"
+    },
+    "./strings": {
+      "import": "./lib/strings/index.js",
+      "require": "./lib/strings/index.js"
+    },
+    "./objects": {
+      "import": "./lib/objects/index.js",
+      "require": "./lib/objects/index.js"
+    }
+  },
   "//": "~~ Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\"."
 }
