+++ src/index.ts
@@ -1,2 +1,2 @@
-export * from "./string";
+export * from "./strings";
 export * from "./objects";
