--- src/objects/__tests__/object-suite.test.ts
@@ -1,5 +1,0 @@
-import { discoverArgumentsTests } from "./discoverArguments.test";
-
-describe("Object helpers", () => {
-  discoverArgumentsTests();
-});
