+++ src/objects/prototypes.ts
@@ -1,0 +1,8 @@
+// import { genericObjectFunction } from "./helpers"
+
+// declare global {
+//     interface Object {
+//         something(): Object
+//         // myObjectFunction: returnType
+//     }
+// }
