--- src/string/__tests__/string-suite.test.ts
@@ -1,9 +1,0 @@
-import { kebabToCamelTests } from "./kebabToCamel.test";
-import { capitalTests } from "./capital.test";
-import { camelToKebabTests } from "./camelToKebab.test";
-
-describe("String helpers", () => {
-  kebabToCamelTests();
-  capitalTests();
-  camelToKebabTests();
-});
