--- src/string/index.ts
@@ -1,3 +1,0 @@
-export * from "./capital";
-export * from "./camelToKebab";
-export * from "./kebabToCamel";
