+++ src/strings/index.ts
